app/requirements.txt
Django==4.0.1
django-debug-toolbar==3.2.4
django-treebeard==4.5.1
django-rest-swagger==2.2.0
drf-spectacular==0.21.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
PyYAML==6.0
